+++ TombEngine/Game/Lara/lara_tests.cpp
@@ -138,6 +138,7 @@ bool TestLaraHang(ItemInfo* item, CollisionInfo* coll)
 	coll->Setup.UpperFloorBound = -STEPUP_HEIGHT;
 	coll->Setup.LowerCeilingBound = 0;
 	coll->Setup.ForwardAngle = lara->Control.MoveAngle;
+	coll->Setup.ForceSolidStatics = true;
 
 	// When Lara is about to move, use larger embed offset for stabilizing diagonal shimmying)
 	int embedOffset = 4;
@@ -925,7 +926,7 @@ bool TestLaraWaterClimbOut(ItemInfo* item, CollisionInfo* coll)
 	int frontFloor = NO_HEIGHT;
 	if (coll->Front.Bridge != NO_VALUE)
 	{
-		auto pointColl = GetPointCollision(*item, item->Pose.Orientation.y, BLOCK(0.5f), -BLOCK(0.5f));
+		auto pointColl = GetPointCollision(*item, item->Pose.Orientation.y, BLOCK(0.2f), -BLOCK(0.5f));
 		frontFloor = (pointColl.GetFloorHeight() - item->Pose.Position.y);
 	}
 	else
@@ -933,7 +934,7 @@ bool TestLaraWaterClimbOut(ItemInfo* item, CollisionInfo* coll)
 		int yOffset = CLICK(1.25f);
 		while (yOffset > -CLICK(2))
 		{
-			auto pointColl = GetPointCollision(*item, item->Pose.Orientation.y, BLOCK(0.5f), yOffset);
+			auto pointColl = GetPointCollision(*item, item->Pose.Orientation.y, BLOCK(0.2f), yOffset);
 
 			frontFloor = pointColl.GetFloorHeight() - item->Pose.Position.y;
 			if (frontFloor > -CLICK(2) &&