+++ TombEngine/Game/control/control.cpp
@@ -452,7 +452,16 @@ void InitializeScripting(int levelIndex, LevelLoadType type)
 	// Run level script if it exists.
 	if (!level.ScriptFileName.empty())
 	{
-		g_GameScript->ExecuteScriptFile(g_GameFlow->GetGameDir() + level.ScriptFileName);
+		auto levelScriptName = g_GameFlow->GetGameDir() + level.ScriptFileName;
+		if (std::filesystem::is_regular_file(levelScriptName))
+		{
+			g_GameScript->ExecuteScriptFile(levelScriptName);
+		}
+		else
+		{
+			TENLog("Level script not found: " + levelScriptName, LogLevel::Warning);
+		}
+
 		g_GameScript->InitCallbacks();
 		g_GameStringsHandler->SetCallbackDrawString([](const std::string& key, D3DCOLOR color, const Vec2& pos, float scale, int flags)
 		{