+++ TombEngine/Game/room.h
@@ -129,6 +129,7 @@ int IsRoomOutside(int x, int y, int z);
 void InitializeNeighborRoomList();
 
 GameBoundingBox& GetBoundsAccurate(const MESH_INFO& mesh, bool getVisibilityBox);
+std::vector<int> GetNeighborRoomNumbers(int roomNumber, unsigned int searchDepth, std::vector<int>& visitedRoomNumbers = std::vector<int>{});
 
 namespace TEN::Collision::Room
 {