+++ CHANGELOG.md
@@ -18,6 +18,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Fixed ghost collision with objects with zero bounds.
 * Fixed several binocular bugs.
 * Fixed faulty death sectors.
+* Fixed shimmy softlocks around static meshes with soft collision.
 * Fixed incorrect climbing out of water on bridge objects and in front of static meshes.
 * Fixed incorrect diving animation when swandiving from a high place.
 * Fixed camera rotating with the player's hips when climbing out of water.
@@ -48,6 +49,7 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
 * Added the ability to change pickup item count by modifying item hit points.
 * Added F12 as alternative to PrtSc for screenshots.
 * Added visible mouse pointer in windowed mode.
+* Added portal debug mode.
 * Added new sound conditions: Quicksand and Underwater.
   - Quicksand - sound effect plays when a moveable is in quicksand.
   - Underwater - sound plays when the camera is submerged.